+++ examples/10-openapi/src/main.ts
@@ -5,7 +5,7 @@ import { AppModule } from './app/app.module';
 new Application()
   .bootstrap(AppModule)
   .then((app) => {
-    app.server.listen(3000);
+    app.server.listen(3000, 'localhost');
   })
   .catch((err) => {
     console.log(err);