+++ packages/openapi/src/swagger-ui/index.ts
@@ -1,6 +1,6 @@
-import SwaggerUI from 'swagger-ui';
+import SwaggerUI = require('swagger-ui');
 import 'swagger-ui/dist/swagger-ui.css';
 
 import { swaggerOptions } from './swagger.config';
 
-SwaggerUI(swaggerOptions.initUi).initOAuth(swaggerOptions.oauthOptions);
+(SwaggerUI as any)(swaggerOptions.initUi).initOAuth(swaggerOptions.oauthOptions);